Misc: Updating XmlGenerator DB model with SDE Inferno 1.0.

+++ Tools/XmlGenerator/StaticData/DgmAttributeTypes.cs
@@ -29,7 +29,7 @@ namespace EVEMon.XmlGenerator.StaticData
         public int? UnitID { get; set; }
 
         [XmlElement("categoryID")]
-        public long? CategoryID { get; set; }
+        public int? CategoryID { get; set; }
 
         [XmlElement("highIsGood")]
         public bool HigherIsBetter { get; set; }
